# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Bonza Bike Tours REST API",
		"description": "REST API description for the Bonza Bike Tours BMS dashboard.",
		"contact": {
			"name": "API Support",
			"email": "alexander.shumilov@toptal.com"
		},
		"version": "1.2"
	},
	"servers": [{
			"url": "bonzabiketours-api.herokuapp.com/api/v1",
			"description": "Heroku server - CI"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3001/api/v1",
			"description": "Staging server"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3002/api/v1",
			"description": "Development server"
		}
	],
	"security": [{
		"api_key": []
	}],
...
	"paths": {
...
		"/purchases/{purchaseId}/payment": {
			"post": {
				"tags": [
					"purchases"
				],
				"security": [{
					"api_key": []
				}],
				"parameters": [{
					"name": "purchaseId",
					"in": "path",
					"description": "Purchase identifier",
					"style": "simple",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"summary": "Creates payment for purchase",
				"requestBody": {
					"description": "Payment description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"paymentDate": {
										"type": "string"
									},
									"amount": {
										"type": "string"
									},
									"method": {
										"type": "string"
									},
									"internalNotes": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Payment was created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/purchases/{purchaseId}/refund": {
			"post": {
				"tags": [
					"purchases"
				],
				"security": [{
					"api_key": []
				}],
				"parameters": [{
					"name": "purchaseId",
					"in": "path",
					"description": "Purchase identifier",
					"style": "simple",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"summary": "Refunds purchase payment",
				"requestBody": {
					"description": "Refund description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"paymentDate": {
										"type": "string"
									},
									"amount": {
										"type": "string"
									},
									"method": {
										"type": "string"
									},
									"internalNotes": {
										"type": "string"
									},
									"refundReason": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Refund was created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/purchases/{purchaseId}/history": {
			"get": {
				"tags": [
					"purchases"
				],
				"security": [{
					"api_key": []
				}],
				"parameters": [{
					"name": "purchaseId",
					"in": "path",
					"description": "Purchase identifier",
					"style": "simple",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"summary": "Purchase history",
				"responses": {
					"200": {
						"$ref": "#/components/responses/PurchaseHistory"
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/purchases/charges/{chargeId}": {
			"put": {
				"tags": [
					"purchases"
				],
				"security": [{
					"api_key": []
				}],
				"parameters": [{
					"name": "chargeId",
					"in": "path",
					"description": "Charge identifier",
					"style": "simple",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"summary": "Updates charge",
				"requestBody": {
					"description": "Application setting description",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"addedToAccounting": {
										"type": "boolean"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Updates charge",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"delete": {
				"tags": [
					"purchases"
				],
				"security": [{
					"api_key": []
				}],
				"parameters": [{
					"name": "chargeId",
					"in": "path",
					"description": "Charge identifier",
					"style": "simple",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"summary": "Deletes charge",
				"responses": {
					"200": {
						"description": "Deletes charge",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
...
		"/emailTemplates": {
			"get": {
				"tags": [
					"emailTemplates"
				],
				"security": [{
					"api_key": []
				}],
				"parameters": [{
					"name": "archived",
					"description": "Specifies if archived email templates are queried as well",
					"in": "query",
					"schema": {
						"type": "string"
					}

				}],
				"summary": "Email templates",
				"responses": {
					"200": {
						"description": "Email templates",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/EmailTemplate"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/emailTemplates/{id}": {
			"put": {
				"tags": [
					"emailTemplates"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Update email template",
				"parameters": [{
					"name": "id",
					"description": "Email template identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"requestBody": {
					"description": "Email description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"templateCode": {
										"type": "string"
									},
									"title": {
										"type": "string"
									},
									"standardText": {
										"type": "string"
									},
									"customerText": {
										"type": "string"
									},
									"tourOperatorText": {
										"type": "string"
									},
									"voucherText": {
										"type": "string"
									},
									"subject": {
										"type": "string"
									},
									"sendFrom": {
										"type": "string"
									},
									"filename": {
										"type": "string"
									},
									"originalFilename": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"delete": {
				"tags": [
					"emailTemplates"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Delete email template",
				"parameters": [{
					"name": "id",
					"description": "Email template identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/emailTemplates/{emailTemplateId}/preview": {
			"get": {
				"tags": [
					"emailTemplates"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Preview email template",
				"parameters": [{
					"name": "emailTemplateId",
					"description": "Email template identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Rendered template",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
...
	},
	"components": {
		"responses": {
			"PurchaseHistory": {
				"description": "Purchase payments and refunds history with overall balance",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"totalNet": {
									"type": "number"
								},
								"history": {
									"type": "array",
									"items": {
										"properties": {
											"chargeID": {
												"type": "number"
											},
											"purchaseID": {
												"type": "number"
											},
											"paymentDate": {
												"type": "string"
											},
											"type": {
												"type": "string"
											},
											"amount": {
												"type": "number"
											},
											"method": {
												"type": "string"
											},
											"internalNotes": {
												"type": "string"
											},
											"refundReason": {
												"type": "string"
											},
											"addedStaff": {
												"type": "number"
											}
										}
									}
								}
							}
						}
					}
				}
			},
			"APISuccessResponse": {
				"description": "Operation was completed successfully",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"message": {
									"type": "string"
								}
							}
						}
					}
				}
			},
...
			"APIUnauthorizedResponse": {
				"description": "Authorize in order to access the resource",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APINotFoundResponse": {
				"description": "Requested resource was not found",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APICreatedEntityResponse": {
				"description": "Returning identifier of created entity",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"id": {
									"type": "integer"
								}
							}
						}
					}
				}
			}
		},
...
		"schemas": {
...
			"EmailTemplate": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
...
		}
	}
}